+++ components/settings/master-data/RiskFactorsTab.tsx
@@ -1,1 +1,0 @@
-export { RiskFactorsIngest as RiskFactorsTab } from "@/components/master-data/ingest/RiskFactorsIngest";